refactor(commands): move sole-caller helpers from lib.rs into their command modules (#233)
Relocate the leftover helpers that are called by exactly one command module and
have no lib.rs unit test out of lib.rs into that module — the #233 helper sweep.
lib.rs drops ~210 lines.

- commands/list.rs    <- spawn_status_check, fetch_plugin_statuses
- commands/add.rs     <- read_persisted_plugin_name
- commands/set.rs     <- read_input_with_esc
- commands/edit.rs    <- hook_indicators, global_hook_indicators, file_with_icon
- commands/profile.rs <- recover_stale_loader_backup
- dropped the now-unused `tokio::task::JoinSet` import from lib.rs

spawn_status_check / fetch_plugin_statuses get `config::Config` ->
`crate::config::Config` to avoid the `commands::config` submodule shadow.

Helpers shared across multiple command modules (resolve_plugin_dst, plural,
prune_*, build_loader_options, ...) stay in lib.rs per the issue. Sole-caller
helpers that are covered by lib.rs unit tests (classify_held_back,
matches_rebuild_filter, the add-lazy / AddTriggerSuggestion cluster,
loader_init_snippet, find_plugin_line_in_toml, ...) stay pinned to lib.rs until
their tests relocate in #224 — moving them now would orphan those tests.

Behaviour-preserving. cargo make check green: 822 tests, plural doctest,
clippy -D warnings --all-targets, fmt, lock-check.

Refs #233, #224, #217.
